requirements.txt
streamlit
pandas
openpyxl
psycopg2-binary
bcrypt
python-dotenv
Pillow
extra-streamlit-components
streamlit-aggrid
